python3-requirements.txt
PyYAML
paho-mqtt
gevent
requests
kubernetes
jsonpath-ng
Mako
